requirements.txt
langchain==0.3.16
langchain-aws==0.2.9
langgraph==0.2.59
langchain-community==0.3.16
boto3==1.35.91
botocore==1.35.91
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.0
newsapi-python==0.2.7
yfinance==0.2.50
arxiv==2.2.0
wikipedia-api==0.8.1
pydantic==2.10.6
httpx==0.28.1
pytest==8.4.2
jupyter==1.1.1
pytest==8.4.2
black==24.10.0
pre-commit==4.3.0
